+++ frontend/lib/data/all_caches.dart
@@ -1,159 +1,3 @@
-import 'package:frontend/models/test_cache.dart';
+import 'package:frontend/models/caches.dart';
 
-final allCaches = [
-  TestCache(
-    name: 'Wooden Mickey Mouse Statue',
-    desc:
-        'The wooden Mickey Mouse statue is located on the fourth floor of the UCF library. The whimsical figure pays homage to the university\'s close proximity to the magic of Walt Disney World. Standing in "The Gallery", the statue serves as both a decorative piece and a reminder of the creativity and innovation that thrive within the university community. Amidst the academic pursuits, this cheerful statue offers a playful nod to the joy and imagination that enriches campus life at UCF.',
-    icon: 'assets/cache_icons/mouse_icon.png',
-    imgUrl: 'assets/cache_images/mouse_image.jpeg',
-    lat: 0,
-    lng: 0,
-    questions: [
-      Question(
-        questionText: 'During what event was the statue donated?',
-        answers: [
-          'UCF’s 25th Anniversary',
-          'UCF’s 30th Anniversary',
-          'UCF’s opening',
-          'UCF’s 10th Anniversary',
-        ],
-      ),
-      Question(
-        questionText: 'What color is the plaque under the statue?',
-        answers: [
-          'Gold',
-          'Brown',
-          'Black',
-          'Silver',
-        ],
-      ),
-      Question(
-        questionText: 'Where does the statue come from?',
-        answers: [
-          'Disney',
-          'The State Governor at the time',
-          'John C. Hitt',
-          'Student Government Association',
-        ],
-      ),
-    ],
-  ),
-  const TestCache(
-    name: 'Koi Pond',
-    desc:
-        'The UCF Koi Pond, nestled within the Burnett Honors College at the University of Central Florida, serves as a tranquil oasis amid the bustling campus life. Tucked away behind the wall of the Burnett Honors College and lush greenery, many students only heard of or seen pictures of the Koi Pond but have never found it right next to the Student Union. The Koi Pond\'s scenic and quiet atmosphere is a geat place to disconnect from the busy schedule and lively activities of the Student Union. ',
-    icon: 'assets/cache_icons/koi_icon.png',
-    imgUrl: 'assets/cache_images/koi_image.jpeg',
-    lat: 0,
-    lng: 0,
-  ),
-  const TestCache(
-    name: 'Foucault Pendulum',
-    desc:
-        'The Foucault pendulum in the Mathematical Sciences Building, named after French physicist Léon Foucault, was part of the original building and has been on campus since October 1970. Back then the pendulum hung unobstructed until people began pulling at the cable that holds the brass ball with sculptured end caps. In an effort to avoid accidents, a see-through barrier was installed. The pendulum broke down in the 1990s and sat idle until Eileen Swindling, ’88, decided to do something about it. The pendulum as restored to working order in the Spring of 2016. Now, the pendulum left many students wondering what caused the pendulum to swing. ',
-    icon: 'assets/cache_icons/pendulum_icon.png',
-    imgUrl: 'assets/cache_images/pendulum_image.jpeg',
-    lat: 0,
-    lng: 0,
-  ),
-  const TestCache(
-    name: 'The Charging Knight',
-    desc:
-        'The Don Reynolds\' statue, or more common "The Charging Knight," located in IOA Plaza just west of FBC Mortgage Stadium. The statue was created in 2007 by artist Don Reynolds out of cast bronze. At the revealing ceremony, former UCF President John C. Hitt noted the statue was representative of UCF’s excellence in academics, its partnerships with the community, its athletics, and UCF\'s push toward national prominence.',
-    icon: 'assets/cache_icons/knight_icon.png',
-    imgUrl: 'assets/cache_images/knight_image.jpeg',
-    lat: 0,
-    lng: 0,
-  ),
-  const TestCache(
-    name: 'The Piano',
-    desc:
-        'The piano is located in the 1st floor of the Teaching Academy building study area. Whether utilized for impromptu performances, student recitals, or simply as a backdrop for scholarly discussions, the piano in the Teaching Academy fosters an environment where passion and learning intertwine. Its presence not only enriches the academic atmosphere but also encourages the exploration of artistic expression within the realm of education. Often in the afternoon, there is alway someone paying, provide a very calming atmosphere.',
-    icon: 'assets/cache_icons/piano_icon.png',
-    imgUrl: 'assets/cache_images/piano_image.jpeg',
-    lat: 0,
-    lng: 0,
-  ),
-  const TestCache(
-    name: 'Memorial for Demetrius Demarcus James Bartholomew III Jr.',
-    desc:
-        'Located in the heart of campus, the dragonfly in the Student Union is a sight many often missed. Nobody know when or how the monument got there. Unfortunately, as mysterious as it appeared, the dragonfly vanished. However traces of the monument still remain for many to make their pilgrimage. ',
-    icon: 'assets/cache_icons/memorial_icon.png',
-    imgUrl: 'assets/cache_images/memorial_image.jpeg',
-    lat: 0,
-    lng: 0,
-  ),
-  const TestCache(
-    name: 'Outdoor View From HEC',
-    desc:
-        'Just a very scenic and relaxing spot on the 4th floor, on top of the staircase of L3Harris Corporation Engineering Center right next to the arboretum. The spot provide a great overlook of the arboretum, campus, and the Bounce Houce. Due to its high, unobstructed view, this spot is great to watch rocket launches from Kennedy Space Center. UCF is SpaceU for a reason!',
-    icon: 'assets/cache_icons/launch_view_icon.png',
-    imgUrl: 'assets/cache_images/launch_view_image.jpeg',
-    lat: 0,
-    lng: 0,
-  ),
-  const TestCache(
-    name: 'Bugs, Rocks, and Nests Collections',
-    desc:
-        'Located in the 1st floor of the Biology building, the various collections is amazing way to learn more about the field of biology while admire stunning artifacts. The hallway features multiple collections from butterflies, beetles, and seashell to rock and ornithology.',
-    icon: 'assets/cache_icons/bug_icon.png',
-    imgUrl: 'assets/cache_images/bug_image.jpeg',
-    lat: 0,
-    lng: 0,
-  ),
-  const TestCache(
-    name: 'Robinson Observatory',
-    desc:
-        'Located right next to the main campus, the Robinson Observatory at UCF stands as a beacon of astronomical exploration and education. It offers students and researchers a window into the mysteries of the cosmos. Equipped with state-of-the-art telescopes and cutting-edge technology, the observatory provides a platform for hands-on learning and discovery. From observing celestial phenomena to engaging in research projects, visitors to the Robinson Observatory are immersed in the wonder of the universe. ',
-    icon: 'assets/cache_icons/observatory_icon.png',
-    imgUrl: 'assets/cache_images/observatory_image.jpeg',
-    lat: 0,
-    lng: 0,
-  ),
-  const TestCache(
-    name: 'Fries/Ketchup sculptures',
-    desc:
-        'Located between the Library, Breezeway, and Trevor Colbourn Hall, the two abstract sculptures often leave students in puzzle. The yellow sculpture is a welded aluminum structure built in 1974, dedicated to John Henry and named Sun Target #1. The red sculpture is a steel structure built in 1975, dedicated to Bruce White and named Alamar. Overtimes, the sculptures has been fondly given the nicknames of Ketchup and Fries by the student body.',
-    icon: 'assets/cache_icons/fries_icon.png',
-    imgUrl: 'assets/cache_images/fries_image.jpeg',
-    lat: 0,
-    lng: 0,
-  ),
-  const TestCache(
-    name: '70s Throwback',
-    desc:
-        'Located in the 2nd floor of Howard Phillips Hall, the “70s Throwback” remind student of the past. The “70s Throwback” act as a time capsule of the early day of campus life and the near by old fashion mailbox complete the package. Even though the phone is not operational, it is still a very cool artifact of the early day of UCF. ',
-    icon: 'assets/cache_icons/old_phone_icon.png',
-    imgUrl: 'assets/cache_images/old_phone_image.jpeg',
-    lat: 0,
-    lng: 0,
-  ),
-  const TestCache(
-    name: 'Old Computer',
-    desc:
-        'Located in the back entrance of between the two Engineering buildings, the display case serve as a nostalgic tribute to the evolution of computing technology. As students and visitors pass by, they are transported back in time, witnessing the progression of hardware and software that has shaped modern computing. The display case not only highlights the advancements in technology but also honors the pioneers and innovators who paved the way for the digital age. ',
-    icon: 'assets/cache_icons/computer_icon.png',
-    imgUrl: 'assets/cache_images/computer_image.jpeg',
-    lat: 0,
-    lng: 0,
-  ),
-  const TestCache(
-    name: 'Cyclorama Sculptures',
-    desc:
-        'Located just outside the School of Performing Arts - Music, the Cyclorama sculptures add to the artistic and creative environment of the side of campus. The Cyclorama sculptures is 6’ high, 6’ wide, 6’ deep and painted bright red. The sculptures is not just for show, student can sit and relax on the sculptures as well. ',
-    icon: 'assets/cache_icons/sculpture_icon.png',
-    imgUrl: 'assets/cache_images/sculpture_image.jpeg',
-    lat: 0,
-    lng: 0,
-  ),
-  const TestCache(
-    name: 'Space Mountain',
-    desc:
-        'Located inside Engineering atrium, the Space Mountain coaster car server as a reminder how UCF can impact the surrounding area. Back in the day, the Walt Disney Company contacted UCF College of Engineering and requested help designing the Space Mountain trains. The park had issues with the restraints and test dummies falling out and contacted UCF who designed the T-Bar restraint system to prevent the incident from happens again. Today, the Space Mountain car stand as a proud example how UCF can help the surrounding community.',
-    icon: 'assets/cache_icons/coaster_icon.png',
-    imgUrl: 'assets/cache_images/coaster_image.jpeg',
-    lat: 0,
-    lng: 0,
-  ),
-];
+Caches? allCaches;
