--- frontend/lib/data/testing_newcaches.dart
@@ -1,4 +1,0 @@
-import "package:flutter/material.dart";
-import 'package:frontend/models/caches.dart';
-
-Caches? newAllCaches;
