+++ frontend/lib/widgets/cachelog/cachelog_screen.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:frontend/models/test_cache.dart';
 import 'package:frontend/widgets/styling/theme.dart';
-import 'package:frontend/data/all_caches.dart';
+import 'package:frontend/data/deprecated_all_caches.dart';
 import 'cache_details.dart';
 import 'caches_grid.dart';
 
@@ -18,7 +18,7 @@ class _CacheLogScreenState extends State<CacheLogScreen> {
   @override
   void initState() {
     super.initState();
-    selectedCache = allCaches[0];
+    selectedCache = deprecatedAllCaches[0];
   }
 
   @override
@@ -43,7 +43,7 @@ class _CacheLogScreenState extends State<CacheLogScreen> {
             Expanded(
               flex: 1,
               child: CachesGrid(
-                allCaches: allCaches,
+                allCaches: deprecatedAllCaches,
                 onCacheSelected: (cache) {
                   setState(() {
                     selectedCache = cache;
